// front-end/config/market/ContellaMarket.types.ts
/**
* This file was automatically generated by @cosmwasm/ts-codegen@0.35.7.
* DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
* and run the @cosmwasm/ts-codegen generate command to regenerate this file.
*/

export interface InstantiateMsg {
  option_address: string;
}
export type ExecuteMsg = {
  list: {
    expires: number;
    id: number;
    price: Uint128;
  };
} | {
  remove_list: {
    id: number;
  };
} | {
  update_price: {
    id: number;
    price: Uint128;
  };
} | {
  set_bid: {
    expires: number;
    id: number;
  };
} | {
  remove_bid: {
    id: number;
  };
} | {
  buy: {
    id: number;
  };
} | {
  accept_bid: {
    bidder: string;
    id: number;
  };
};
export type Uint128 = string;
export type QueryMsg = {
  contract_params: {};
} | {
  list_items: {};
} | {
  list_items_page: {
    amount: number;
    key: number;
  };
} | {
  get_list_items_byid: {
    id: number;
  };
} | {
  owner_list_items: {
    addr: string;
  };
} | {
  owner_un_list_items: {
    addr: string;
  };
} | {
  bid_list: {
    id: number;
  };
};
export type Addr = string;
export type Timestamp = Uint64;
export type Uint64 = string;
export type ArrayOfTupleOfAddrAndBid = [Addr, Bid][];
export interface Bid {
  expires_at: Timestamp;
  price: Uint128;
}
export interface ContractParams {
  onwer_address: Addr;
  option_address: Addr;
}
export interface ListItemData {
  collateral: Coin;
  counter_offer: Coin;
  creator: Addr;
  expires: Timestamp;
  list_expires: Timestamp;
  owner: Addr;
  price: Uint128;
}
export interface Coin {
  amount: Uint128;
  denom: string;
  [k: string]: unknown;
}
export type ArrayOfTupleOfUint64AndListItemData = [number, ListItemData][];